reference code for retrivals agents/flusso_vision_agent/requirements.txt
# Flusso Vision Agent - Requirements
torch>=2.0.0
open-clip-torch>=2.20.0
pinecone>=5.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
colorlog>=6.7.0
numpy>=1.24.0
